dplyr::mutate instead of just mutate
That last commit fixed that issue. Now it is crying because it cannot find the 'mutate' function. We will see if putting "dplyr::" before will help. This is good to know, apparently in order to run this stuff through python you need to specify the library that the function comes from.