Including Null values in XWIDTH Count_Calc
One line 73 I put some comments. XWIDTH was not matching for the station 404M07362. the legacy calculator arrived at the number that it did by dividing by 21 (Total number of observations, including Nulls.) The R package rather, divided by 19, which would be the number of non null observations. I suspect that the R package is correct in the calculation of this metric, while the legacy calculator was off.