Adding extra rounding to XBKF_H
I rounded the sum of the result values of XBKF_H before doing the division by the count in order to see if this would match the numbers for station LALT500. This Metric calculation matched with all stations except that one, and to be honest, I do think that the R package is more accurate in this case.