Completely re did the code for Percentiles
So, although we believe the code for calculation of Particle Size and Particulate Particle Size percentiles was correct on our end, we found 
the numbers did not match what was coming out from the original VBA code. After studying the VBA code, we found that it did not calculate the distribution of the particle sizes, but rather, sorted the values and grabbed certain values at the corresponding indices. For example, to get the 90th percentile of a set of 80 values, it would do 0.9 x 80 = 72 and so grab the 72nd value in the sorted list.