Fixing Counts for Sub Size Class
I am seeing if applying this will fix the issue with the counts. In the past we assumed that the old calculation method was not considering NA's except for station 404M07362, but I am now questioning that. I am experimenting to see how the old method is coming up with the counts for this one. We will have to test the part where the counts are assigned to variables and reported, because I fear that they may possibly be assigned out of order.