Fixing subsetting for H_SubNat
I think we goofed. Looks like we were trying to filter out the Not recorded's, but unsuccessfully, since the VariableResult column got transformed to uppercase. 

By the way, transforming it to uppercase was a mechanism to deal with messed up data. We need lookup list checks on that variable result column